Fix for --incompatible_remove_old_python_version_api
This is a one-liner to rename `default_python_version` for a `py_binary` target.

This does not rename the corresponding attribute on the `parfile` rule. That attribute is currently marked mandatory, though it doesn't appear to be used. If we want to keep it, and if we want a gentle migration period, we'd need to introduce a new `python_version` attribute for `parfile`, make them both non-mandatory, and validate them at anaylsis time, similarly to how Bazel handled the rename of this attribute for `py_binary`/`py_test`.